# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |
    This is the API documentation for WASAText.
  version: 2.1.0
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
...
paths:
  /session:
    post:
      tags:
        - Users
      operationId: doLogin
      summary: Registers or logs in a user
      description: |
        If the user does not exist, it will be registered and an identifier will be created.
        If the user exists, the user profile is returned.
      security: []
      requestBody:
        required: true 
        content:
          application/json:
            schema:
              description: Payload to login/register the user.
              type: object
              properties:
                username: { $ref: "#/components/schemas/Username" }
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                description: Response schema for user login.
                properties:
                  profile: { $ref: "#/components/schemas/User"}
              example:
                id: 17
                profile:
                  username: "alice1"
                  pic: "iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg=="
        '201':
          description: Successful registration
          content:
            application/json:
              schema:
                type: object
                description: Response schema for user registration.
                properties:
                  profile: { $ref: "#/components/schemas/User"}
              example:
                id: 17
                profile:
                  username: "alice1"
                  pic: "iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg=="
        '400': { $ref: "#/components/responses/BadRequest" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  /users:
    get:
      tags:
        - Users
      operationId: searchUser
      summary: Search for a user by username
      description: Searches for user matching a query string
      security:
        - bearerAuth: []
      parameters:
      - name: searched_user
        description: Username to search.
        in: query
        required: true
        schema: { $ref: "#/components/schemas/Username" }
      
      responses:
        "200":
          description: List of usernames that matched with the searched one
          content:
            application/json:
              schema:
                type: array
                description: Array of users.
                minItems: 0
                maxItems: 700
                items: { $ref: "#/components/schemas/User" }
              example:
                - username: "alice1"
                  pic: "iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg=="
                - username: "alice2"
                  pic: "iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg=="
                
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '404': { $ref: "#/components/responses/NotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  /users/me/username:
    put:
      tags:
        - Users
      operationId: setMyUserName
      summary: Updates the current user's username
      description: Updates the user's username and returns the updated user profile
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              description: Payload to update the user's name
              type: object
              properties:
                username: { $ref: "#/components/schemas/Username" }
        required: true
      responses:
        '200':
          description: Username updated successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User"}
              example:
                  username: "alice3"
                  pic: "iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg=="
        '400': { $ref: "#/components/responses/BadRequest" }
        '404': { $ref: "#/components/responses/NotFound" }
  /users/me/pic:
    put:
      tags:
        - Users
      operationId: setMyPhoto
      summary: Uplaods or updates the current user's profile picture
      description: Uploads or updates a user's profile picture and return the updated profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Payload to update the user's profile picture
              type: object
              properties:
                pic: { $ref: "#/components/schemas/Pic" }
              required: [pic]
      responses:
        '200':
          description: Profile pic updated successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
              example:
                  username: "alice1"
                  pic: "iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg=="
        '400': { $ref: "#/components/responses/BadRequest" }
        '404': { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Username:
      description: Name of a user
      type: string
      pattern: '^[a-zA-Z0-9_.-]{3,13}$'
      minLength: 3
      maxLength: 25
      example: "alice123"
...
    User:
      description: User's profile, which is username and profile picture
      type: object
      properties:
        username: { $ref: "#/components/schemas/Username" }
        pic: { $ref: "#/components/schemas/Pic" }
    Pic:
      description: Base64-encoded image data, it can be JPEG, PNG, GIF, WebP
      type: string
      format: byte
      pattern: '^[A-Za-z0-9+/]+={0,2}$'
      minLength: 100
      maxLength: 500000
      example: "iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg=="
...
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (e.g., missing or invalid fields).
    Unauthorized:
      description: The request requires user authentication or the provided credentials are invalid.
    NotFound:
      description: The requested resource was not found.
    InternalServerError:
      description: The server encountered an internal error. Check server logs for more details.